added tests, moved the function to RegionLogic.ts
