+++ src/components/ExposedHelpButton.tsx
@@ -1,25 +1,10 @@
 import React, {useCallback} from 'react';
 import {Box, ButtonSingleLine, ErrorBox} from 'components';
-import {isRegionActive} from 'shared/RegionLogic';
+import {getExposedHelpURL, isRegionActive} from 'shared/RegionLogic';
 import {useRegionalI18n} from 'locale/regional';
 import {useStorage} from 'services/StorageService';
 import {Linking} from 'react-native';
 import {captureException} from 'shared/log';
-import {Region} from 'shared/Region';
-
-export function getExposedHelpURL(region: Region | undefined, regionActive: boolean, regionalI18n: any) {
-  const nationalURL = regionalI18n.translate(`RegionContent.ExposureView.Inactive.CA.URL`);
-  if (region !== undefined && region !== 'None') {
-    const regionalURL = regionActive
-      ? regionalI18n.translate(`RegionContent.ExposureView.Active.${region}.URL`)
-      : regionalI18n.translate(`RegionContent.ExposureView.Inactive.${region}.URL`);
-    if (regionalURL === '') {
-      return nationalURL;
-    }
-    return regionalURL;
-  }
-  return nationalURL;
-}
 
 export const ExposedHelpButton = () => {
   const {region} = useStorage();
@@ -38,10 +23,10 @@ export const ExposedHelpButton = () => {
 
   const cta = getCTA();
   const onPress = useCallback(() => {
-    Linking.openURL(getExposedHelpURL(region, regionActive, regionalI18n)).catch(error =>
+    Linking.openURL(getExposedHelpURL(region, regionalI18n)).catch(error =>
       captureException('An error occurred', error),
     );
-  }, [region, regionActive, regionalI18n]);
+  }, [region, regionalI18n]);
 
   if (cta === '') {
     return <ErrorBox marginTop="m" />;
